Title: Ｍax number of guesses to win and next steps.
Desription:
Now you can only guess up to 15 times or else the game is over! Next I want to 1.make an option to quit the game. 2.Print that a guess other than a number was not counted, AND 3.Print That a guess which is not a number was counted. 4.I plan to make levels of difficulty starting with 20 guesses all the way down to one or until user quits.